Cleaned up files
Moved distance function to utils folder, removed unused imports, and removed charity export.
