+++ src/app/(tabs)/donate.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import * as Location from 'expo-location';
 import { useRouter } from 'expo-router';
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Alert, StyleSheet, View } from 'react-native';
 import { Button, IconButton, Text, TextInput } from 'react-native-paper';
 
@@ -12,28 +12,6 @@ export default function FindCharityScreen() {
   const [searchText, setSearchText] = useState('');
   const [useLocation, setUseLocation] = useState(false);
   const router = useRouter();
-
-  // const searchNearbyCharities = async (lat: number, lon: number) => {
-  //   console.log(`Searching for charities near (${lat}, ${lon})...`);
-  //   // Replace with backend or public charity API call
-  //   // const response = await fetch(`https://api.example.com/charities?lat=${lat}&lon=${lon}`);
-  //   // const results = await response.json();
-  // };
-
-  // Calculates the distance between two coordinates using the curvature of the Earth
-  const calculateDistance = (lat1: number, lon1: number, lat2: number, lon2: number): number => {
-    const R = 3959; // Earth's radius in miles
-    const dLat = ((lat2 - lat1) * Math.PI) / 180;
-    const dLon = ((lon2 - lon1) * Math.PI) / 180;
-    const a =
-      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-      Math.cos((lat1 * Math.PI) / 180) *
-        Math.cos((lat2 * Math.PI) / 180) *
-        Math.sin(dLon / 2) *
-        Math.sin(dLon / 2);
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-    return R * c;
-  };
 
   const handleSearchPress = () => {
     Alert.alert(
