+++ src/stores/charities.ts
@@ -35,5 +35,3 @@ export const getCharity = (cid: string): Charity | undefined => {
 export const setCharity = (cid: string, charity: Charity): void => {
   charitiesStore.set(cid, charity);
 };
-
-export { Charity };
